requirement.txt
Django>=4.2
djangorestframework>=3.14

ultralytics>=8.3.0
opencv-python>=4.7.0

numpy>=1.22
requests>=2.28
PyYAML>=6.0
pillow>=9.5

reportlab>=3.6

Flask==3.1.2
pyserial==3.5

torch>=2.0.0
torchvision>=0.15.0

paho-mqtt>=1.6.1

psutil>=5.9.0
